chore: Ignore temporary files in .gitignore and add recipe search endpoint
Ignoring temporary files in the `.gitignore` file and adding the recipe search endpoint to the `recipe_router.go` file. Also, renaming the `RecipeIngredientDTO` struct to `CreateRecipeIngredientDTO` in the `recipe_ingredient_dto.go` file to match the purpose of the struct. Additionally, refactoring the `RecipeTagFromDTO` function to `RecipeTagFromIds` in the `recipe_tag.go` file for improved clarity. Lastly, updating the `Save` method in the `RecipeTag` and `RecipeStep` models to return only an `error` instead of the entire struct.